requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
asyncio-dgram==2.2.0
attrs==24.2.0
Brotli==1.1.0
cffi==1.17.1
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
mcstatus==11.1.1
msgspec==0.18.6
multidict==6.1.0
propcache==0.2.1
py-cord==2.6.1
pycares==4.5.0
pycparser==2.22
python-dotenv==1.0.1
yarl==1.18.3
